+++ result
@@ -1,0 +1,1 @@
+/nix/store/gl0hz8fgvms0fw3xmazghfvks7krwsd3-nix-mate.tar.gz
