+++ src/NixMate/Actions/Search.hs
@@ -1,7 +1,5 @@
-module Actions.Search (search, displayPackage, safeShell) where
+module NixMate.Actions.Search (search, displayPackage, safeShell) where
 
-import qualified Actions.CreateNixFile as Actions
-import Control.Exception (try)
 import qualified Data.Aeson as JSON
 import Data.Bifunctor (first)
 import Data.ByteString.Lazy.Char8 as Char8
@@ -10,25 +8,15 @@ import Data.Coerce
 import qualified Data.List as L
 import qualified Data.Map as M
 import Data.Maybe (fromMaybe)
-import System.Process
-import Types.Config
-import Types.Search
+import qualified NixMate.Actions.CreateNixFile as Actions
+import NixMate.Shared
+import NixMate.Types.Config
+import NixMate.Types.Search
 
 -- we have two methods of searching
 -- nix search <package> - which is cached nicely but does not look deep into
 -- stuff like haskellPackages
 -- it would not find Agda, for instance, which is haskellPackages.Agda
-
--- run a shell action that throws
-safeShell :: String -> String -> IO String
-safeShell command def = do
-  let myShell = (shell command) {cwd = Just "."}
-  either' <- try (readCreateProcess myShell "")
-  case (either' :: Either IOError String) of
-    Right a ->
-      pure a
-    _ ->
-      pure def
 
 -- -qa --description
 searchDescription :: Dependency -> String
@@ -68,7 +56,7 @@ search ::
   IO (Either SearchError [(Dependency, SearchPackage)])
 search cfg depName = do
   Actions.savePackageNix cfg
-  str <- safeShell (searchDescription depName) ""
+  str <- safeShellWithDefault (searchDescription depName) ""
   case decodeFromString str of
     Just items -> pure (findMatch items)
     _ -> pure (Left CouldNotReadJson)
