+++ src/NixMate/Types/Config.hs
@@ -3,7 +3,7 @@
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 
-module Types.Config where
+module NixMate.Types.Config where
 
 import qualified Data.Aeson as JSON
 import Data.Set (Set)
@@ -11,26 +11,58 @@ import GHC.Generics
 
 -- which commit hash of nixpkgs are we pinned to?
 newtype Rev = Rev String
-  deriving newtype (Eq, Ord, Show, JSON.FromJSON, JSON.ToJSON, Semigroup, Monoid)
+  deriving newtype
+    ( Eq,
+      Ord,
+      Show,
+      JSON.FromJSON,
+      JSON.ToJSON,
+      Semigroup,
+      Monoid
+    )
 
 displayRev :: Rev -> String
 displayRev (Rev s) = s
 
 -- what is the hash of that commit?
 newtype Sha256 = Sha256 String
-  deriving newtype (Eq, Ord, Show, JSON.FromJSON, JSON.ToJSON)
+  deriving newtype
+    ( Eq,
+      Ord,
+      Show,
+      JSON.FromJSON,
+      JSON.ToJSON
+    )
 
 -- what is the name of our project?
 newtype ProjectName = ProjectName String
-  deriving newtype (Eq, Ord, Show, JSON.FromJSON, JSON.ToJSON)
+  deriving newtype
+    ( Eq,
+      Ord,
+      Show,
+      JSON.FromJSON,
+      JSON.ToJSON
+    )
 
 -- a dependency we have
 newtype Dependency = Dependency String
-  deriving newtype (Eq, Ord, Show, JSON.FromJSON, JSON.ToJSON)
+  deriving newtype
+    ( Eq,
+      Ord,
+      Show,
+      JSON.FromJSON,
+      JSON.ToJSON
+    )
 
 -- where to output nix derivation
 newtype ShellPath = ShellPath String
-  deriving newtype (Eq, Ord, Show, JSON.FromJSON, JSON.ToJSON)
+  deriving newtype
+    ( Eq,
+      Ord,
+      Show,
+      JSON.FromJSON,
+      JSON.ToJSON
+    )
 
 -- our config file in the root of the project
 data Config = Config
@@ -44,3 +76,4 @@ data Config = Config
 
 data ConfigError
   = CouldNotLoadConfig
+  deriving (Eq, Ord, Show)
