+++ src/NixMate/Types/CreateNixFile.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 
-module Types.CreateNixFile where
+module NixMate.Types.CreateNixFile where
 
 newtype Derivation = Derivation String
   deriving newtype (Show, Semigroup)
